requirements.txt
opencv-python
easyocr
pytesseract
pillow
keyboard
mouse
mss
pyautogui
pyscreenshot
PyQt5